+++ geo_activity_playground/core/activity_parsers.py
@@ -212,9 +212,13 @@ def read_gpx_activity(path: pathlib.Path, open) -> pd.DataFrame:
                     time = dateutil.parser.parse(str(point.time))
                 assert isinstance(time, datetime.datetime)
                 time = time.astimezone(datetime.timezone.utc)
-                points.append((time, point.latitude, point.longitude))
+                points.append((time, point.latitude, point.longitude, point.elevation))
 
-    return pd.DataFrame(points, columns=["time", "latitude", "longitude"])
+    df = pd.DataFrame(points, columns=["time", "latitude", "longitude", "altitude"])
+    # Some files don't have altitude information. In these cases we remove the column.
+    if not df["altitude"].any():
+        del df["altitude"]
+    return df
 
 
 def read_tcx_activity(path: pathlib.Path, opener) -> pd.DataFrame: